requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.1
attrs==23.1.0
blinker==1.7.0
Cerberus==1.3.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.7
distlib==0.3.7
docopt==0.6.2
fastapi==0.104.1
Flask==3.0.0
frozenlist==1.4.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
isort==5.13.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
packaging==23.2
pep517==0.13.1
pip-api==0.0.30
pipreqs==0.4.13
platformdirs==4.1.0
plette==0.4.4
pluggy==1.3.0
psycopg2==2.9.9
py==1.11.0
pydantic==2.5.2
pydantic_core==2.14.5
pylint==3.0.2
pytest==6.2.5
python-dotenv==1.0.0
requests==2.31.0
requirementslib==3.0.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
tomlkit==0.12.3
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
Werkzeug==3.0.1
yarg==0.1.9
yarl==1.9.3
